OpenDCL 7.0 Alpha 7 - Added support for Bricscad V13, implemented several feature requests
git-svn-id: https://svn.code.sf.net/p/opendcl/code/Trunk@287 07a8c159-f786-4120-a571-e88efff3e5ae

+++ Studio/TabsPane.cpp
@@ -122,7 +122,7 @@ void CTabsPane::Setup()
 	{
 		TDclFormPtr pForm = (*iter);
 		CTabInfo *pTab = new CTabInfo(pForm);
-		short idxTab = pForm->GetTabIndex();
+		unsigned short idxTab = (unsigned short)pForm->GetTabIndex();
 		if (idxTab < m_pTabCaptions->size())
 			pTab->msCaption = m_pTabCaptions->GetConstStringArrayPtr()->at( idxTab );
 		if( pTab->msCaption.IsEmpty() )
